+++ tests/test_script.py
@@ -44,8 +44,6 @@ def test_lineapy_init_with_options():
             temp_dir_name,
             "--do-not-track=true",
             "--logging-level=debug",
-            "--mlflow-tracking-uri=sqlite://",
-            "--default-ml-models-storage-backend=mlflow",
             "init",
         ]
     )
@@ -57,8 +55,6 @@ def test_lineapy_init_with_options():
     assert generated_config["home_dir"] == temp_dir_name
     assert generated_config["do_not_track"] == "True"
     assert generated_config["logging_level"] == "DEBUG"
-    assert generated_config["mlflow_tracking_uri"] == "sqlite://"
-    assert generated_config["default_ml_models_storage_backend"] == "mlflow"
 
 
 @pytest.mark.slow
